uvi working and print out order changed
The uvi readout from the si1133 sensor is stored in the ir field the light sensor struct.
Flags to indicate that lux and ir (the uvi reading from sensor) is set to one when new data is
written in the fields  of lux and ir. The order of print out in the console has been changed so lux is printed first and ir (uvi) second